Added more material on parallel and data-driven testing for JUnit

+++ src/samples/junit-quick-start/src/main/java/net/serenitybdd/samples/junit/model/FrequentFlyer.java
@@ -40,7 +40,7 @@ public class FrequentFlyer {
         }
 
         public void kilometers() {
-            incrementBalanceBy(distance / 10);
+            incrementBalanceBy((int) (((float)distance) / 10.0));
             updateStatus();
         }
     }
